me.awesomemoder316.* -> io.github.awesomemoder316.*
