+++ src/main/java/io/github/awesomemoder316/passwordlogon/MessageConfig.java
@@ -1,4 +1,4 @@
-package me.awesomemoder316.passwordlogon;
+package io.github.awesomemoder316.passwordlogon;
 
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
