+++ src/main/java/io/github/awesomemoder316/passwordlogon/listeners/OnPlayerLeave.java
@@ -1,14 +1,15 @@
-package me.awesomemoder316.passwordlogon.listeners;
+package io.github.awesomemoder316.passwordlogon.listeners;
 
-import me.awesomemoder316.passwordlogon.Utils;
+import io.github.awesomemoder316.passwordlogon.Utils;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerQuitEvent;
 
 public class OnPlayerLeave implements Listener {
 
-    @EventHandler
+    @EventHandler(priority = EventPriority.HIGHEST)
     public void teleportBack(PlayerQuitEvent e) {
         Player p = e.getPlayer();
