+++ src/main/resources/config.yml
@@ -3,7 +3,6 @@
 # y is the level at which the platform is at. y, and 2 blocks above y, will be cleared in order to stop the player from suffocating.
 # The area between z1 and z2 is the z position of the platform (inclusive of z1 and z2)
 
-configVersion: 1.3.1
 overworld:
   x1: -2.0
   x2: 2.0
